consumer/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
colorama==0.4.4
constantly==15.1.0
cors==1.0.1
cryptography==36.0.2
daphne==3.0.2
Django==4.0.4
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djongo==1.3.6
filelock==3.6.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==3.12.3
pyOpenSSL==22.0.0
PySocks==1.7.1
pika==1.2.1
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
tldextract==3.2.1
Twisted==22.4.0
#twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
websocket==0.2.1
websocket-client==1.3.2
zope.event==4.5.0
zope.interface==5.4.0
